+++ packages/core/src/indexing-store/realtime.ts
@@ -11,7 +11,12 @@ import {
 import type { IndexingErrorHandler, SchemaBuild } from "@/internal/types.js";
 import { prettyPrint } from "@/utils/print.js";
 import { startClock } from "@/utils/timer.js";
-import { type QueryWithTypings, type Table, getTableName } from "drizzle-orm";
+import {
+  type QueryWithTypings,
+  type Table,
+  getTableName,
+  isTable,
+} from "drizzle-orm";
 import { drizzle } from "drizzle-orm/pg-proxy";
 import {
   type IndexingStore,
@@ -19,7 +24,7 @@ import {
   checkTableAccess,
   validateUpdateSet,
 } from "./index.js";
-import { getCacheKey, getWhereCondition } from "./utils.js";
+import { getCacheKey, getPrimaryKeyCache, getWhereCondition } from "./utils.js";
 
 export const createRealtimeIndexingStore = ({
   common,
@@ -34,6 +39,9 @@ export const createRealtimeIndexingStore = ({
 }): IndexingStore => {
   let qb: QB = undefined!;
   let isProcessingEvents = true;
+
+  const tables = Object.values(schema).filter(isTable);
+  const primaryKeyCache = getPrimaryKeyCache(tables);
 
   const storeMethodWrapper = (fn: (...args: any[]) => Promise<any>) => {
     return async (...args: any[]) => {
@@ -115,8 +123,8 @@ export const createRealtimeIndexingStore = ({
 
                 for (let i = 0; i < values.length; i++) {
                   if (
-                    getCacheKey(table, values[i]) ===
-                    getCacheKey(table, result[resultIndex]!)
+                    getCacheKey(table, values[i], primaryKeyCache) ===
+                    getCacheKey(table, result[resultIndex]!, primaryKeyCache)
                   ) {
                     rows.push(result[resultIndex++]!);
                   } else {
@@ -161,9 +169,8 @@ export const createRealtimeIndexingStore = ({
 
                   if (row) {
                     const set =
-                      typeof valuesU === "function"
-                        ? validateUpdateSet(table, valuesU(row), row)
-                        : validateUpdateSet(table, valuesU, row);
+                      typeof valuesU === "function" ? valuesU(row) : valuesU;
+                    validateUpdateSet(table, set, row, primaryKeyCache);
 
                     rows.push(
                       await qb.wrap((db) =>
@@ -194,9 +201,8 @@ export const createRealtimeIndexingStore = ({
 
                 if (row) {
                   const set =
-                    typeof valuesU === "function"
-                      ? validateUpdateSet(table, valuesU(row), row)
-                      : validateUpdateSet(table, valuesU, row);
+                    typeof valuesU === "function" ? valuesU(row) : valuesU;
+                  validateUpdateSet(table, set, row, primaryKeyCache);
 
                   return qb.wrap((db) =>
                     db
@@ -307,7 +313,9 @@ export const createRealtimeIndexingStore = ({
               throw error;
             }
 
-            const set = validateUpdateSet(table, values(row), row);
+            const set = values(row);
+            validateUpdateSet(table, set, row, primaryKeyCache);
+
             return qb.wrap((db) =>
               db
                 .update(table)
@@ -317,7 +325,8 @@ export const createRealtimeIndexingStore = ({
                 .then((res) => res[0]),
             );
           } else {
-            const set = validateUpdateSet(table, values, row!);
+            const set = values;
+            validateUpdateSet(table, set, row!, primaryKeyCache);
             return qb.wrap((db) =>
               db
                 .update(table)