+++ packages/core/src/indexing/index.ts
@@ -220,9 +220,11 @@ export const createIndexing = ({
     }
   };
 
-  const executeEvent = async ({ event }: { event: Event }): Promise<void> => {
+  // metric label for "ponder_indexing_function_duration"
+  const metricLabel: { event: string } = { event: "" };
+  const executeEvent = async (event: Event): Promise<void> => {
     const indexingFunction = indexingFunctions[event.name];
-    const metricLabel = { event: event.name };
+    metricLabel.event = event.name;
 
     try {
       context.chain.id = event.chainId;
@@ -345,7 +347,7 @@ export const createIndexing = ({
           msg: `Started indexing function (event="${event.name}", checkpoint=${event.checkpoint})`,
         });
 
-        await executeEvent({ event });
+        await executeEvent(event);
 
         common.logger.trace({
           service: "indexing",
@@ -353,7 +355,6 @@ export const createIndexing = ({
         });
       }
 
-      // set completed events
       updateCompletedEvents();
     },
   };