+++ packages/react/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ponder/react
 
+## 0.12.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.20
+
+## 0.12.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.19
+
+## 0.12.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.18
+
+## 0.12.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.17
+
 ## 0.12.16
 
 ### Patch Changes